Revert "Move into nighthawk-benchmark"
This reverts commit fb28c1f41df935ed17aa3d7db5a53c75bf8446d6.
