--- perf/nighthawk-benchmark/Chart.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-name: benchmark
-version: 1.0
-description: Helm chart for istio twoPodTest
-keywords:
-  - istio
-  - performance
-sources:
-  - http://github.com/istio/istio
-engine: gotpl
-icon: https://istio.io/favicons/android-192x192.png
