--- perf/nighthawk-benchmark/configs/istio/mixer_latency.yaml
@@ -1,21 +1,0 @@
-# Configuration Set2: Latency Quantiles with mixer enabled
-telemetry_mode: "mixer"
-conn:
-    - 2
-    - 4
-    - 8
-    - 16
-    - 32
-    - 64
-qps:
-    - 1000
-duration: 240
-metrics:
-    - p50
-    - p90
-    - p99
-perf_record: true
-run_bothsidecar: true
-run_serversidecar: true
-run_clientsidecar: true
-run_baseline: true
