--- perf/nighthawk-benchmark/configs/istio/none_cpu_mem.yaml
@@ -1,20 +1,0 @@
-# Configuration Set3: CPU and memory with mixer disabled
-telemetry_mode: "none"
-conn:
-    - 16
-qps:
-    - 10
-    - 100
-    - 500
-    - 1000
-    - 2000
-    - 3000
-duration: 240
-metrics:
-    - cpu
-    - mem
-perf_record: false
-run_bothsidecar: true
-run_serversidecar: true
-run_clientsidecar: true
-run_baseline: true
