--- perf/nighthawk-benchmark/configs/istio/none_latency.yaml
@@ -1,22 +1,0 @@
-# Configuration Set4: Latency Quantiles with mixer disabled
-telemetry_mode: "none"
-conn:
-    - 2
-    - 4
-    - 8
-    - 16
-    - 32
-    - 64
-qps:
-    - 1000
-duration: 240
-size: 1024
-metrics:
-    - p50
-    - p90
-    - p99
-perf_record: true
-run_bothsidecar: true
-run_serversidecar: true
-run_clientsidecar: true
-run_baseline: true
