--- perf/nighthawk-benchmark/configs/istio/plaintex_cpu_mem.yaml
@@ -1,22 +1,0 @@
-telemetry_mode: "none"
-conn:
-    - 16
-qps:
-    - 10
-    - 100
-    - 500
-    - 1000
-    - 2000
-    - 3000
-duration: 240
-size: 1024
-metrics:
-    - cpu
-    - mem
-perf_record: false
-run_bothsidecar: true
-run_serversidecar: false
-run_clientsidecar: false
-run_baseline: false
-
-extra_labels: "plaintext"
