--- perf/nighthawk-benchmark/configs/istio/plaintext_latency.yaml
@@ -1,24 +1,0 @@
-telemetry_mode: "none"
-conn:
-    - 2
-    - 4
-    - 8
-    - 16
-    - 32
-    - 64
-qps:
-    - 1000
-duration: 240
-metrics:
-    - p50
-    - p90
-    - p99
-perf_record: true
-# we only care about both sidecar mode, which is the base line for mTLs overhead comparision.
-run_bothsidecar: true
-run_serversidecar: false
-run_clientsidecar: false
-run_baseline: false
-
-extra_labels: "plaintext"
-
