--- perf/nighthawk-benchmark/configs/istio/telemetryv2_cpu_mem.yaml
@@ -1,20 +1,0 @@
-# Configuration Set5: CPU and memory with telemetryv2 using NullVM.
-telemetry_mode: "v2-nullvm"
-conn:
-    - 16
-qps:
-    - 10
-    - 100
-    - 500
-    - 1000
-    - 2000
-    - 3000
-duration: 240
-metrics:
-    - cpu
-    - mem
-perf_record: false
-run_bothsidecar: true
-run_serversidecar: true
-run_clientsidecar: true
-run_baseline: true
