--- perf/nighthawk-benchmark/configs/istio/telemetryv2_latency.yaml
@@ -1,21 +1,0 @@
-# Configuration Set6: Latency Quantiles with telemetry v2 using NullVM.
-telemetry_mode: "v2-nullvm"
-conn:
-    - 2
-    - 4
-    - 8
-    - 16
-    - 32
-    - 64
-qps:
-    - 1000
-duration: 240
-metrics:
-    - p50
-    - p90
-    - p99
-perf_record: true
-run_bothsidecar: true
-run_serversidecar: true
-run_clientsidecar: true
-run_baseline: true
