--- perf/nighthawk-benchmark/configs/linkerd/linkerd_latency.yaml
@@ -1,18 +1,0 @@
-mesh: "linkerd"
-conn:
-    - 2
-    - 4
-    - 8
-    - 16
-    - 32
-    - 64
-qps:
-    - 1000
-duration: 240
-metrics:
-    - p50
-    - p90
-    - p99
-perf_record: false
-run_serversidecar: true
-run_baseline: true
