--- perf/nighthawk-benchmark/flame/.gitignore
@@ -1,3 +1,0 @@
-FlameGraph
-*.perf
-*.data
