--- perf/nighthawk-benchmark/runner/Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-requests = "*"
-pytz = "*"
-bokeh = "*"
-pandas = "==0.24.2"
-numpy = "*"
-pyyaml = "*"
-
