--- perf/nighthawk-benchmark/templates/mtls.yaml
@@ -1,28 +1,0 @@
-apiVersion: authentication.istio.io/v1alpha1
-kind: Policy
-metadata:
-  name: "default"
-  namespace: {{ $.Values.namespace }}
-spec:
-  peers:
-  - mtls:
-      mode: PERMISSIVE
----
-apiVersion: networking.istio.io/v1alpha3
-kind: DestinationRule
-metadata:
-  name: fortioserver
-  namespace: {{ $.Values.namespace }}
-spec:
-  host:  fortioserver
-  trafficPolicy:
-    portLevelSettings:
-    - port:
-        number: 8080
-      tls:
-        mode: ISTIO_MUTUAL
-    - port:
-        number: 8077
-      tls:
-        mode: DISABLE
-
