--- perf/nighthawk-benchmark/values.yaml
@@ -1,46 +1,0 @@
-# for client side setting, IP Ranges is server side service or pod
-excludeOutboundIPRanges: ""
-includeOutboundIPRanges: ""
-
-appresources: {}
-
-proxy:
-  cpu: "1000m"
-  memory: "512m"
-
-appresources1:
-  requests:
-    cpu: "800m"
-    memory: "1000m"
-
-fortioImage: fortio/fortio:latest
-curlImage: tutum/curl:trusty
-nighthawkImage: envoyproxy/nighthawk-dev:latest
-domain: local
-gateway: fortio-gateway
-rbac:
-  # If true, generates random RBAC policies for the fortio service.
-  enabled: false
-  # Number of RBAC policies generated.
-  # Note: The generated policies will have a random name. One of the policies allows everyone to access fortioserver and
-  # the other policies allow a random non exist user to access it.
-  numPolicies: 3
-
-server:  # server overrides
-  #tlsmode: DISABLE
-  tlsmode: ISTIO_MUTUAL
-  expose: false
-  inject: "false"
-  injectL: "disabled" # "enabled" or "disabled"
-
-client: # client overrides
-  #tlsmode: DISABLE
-  tlsmode: ISTIO_MUTUAL
-  expose: true
-  inject: "false"
-  injectL: "disabled" # "enabled" or "disabled"
-
-cert: false
-interceptionMode: REDIRECT
-
-namespace: ""
